Move Metric.Registry embedded schemas to a different file
